NLP/requirements.txt
elevenlabs>=1.50.0
streamlit>=1.30.0
python-dotenv>=1.0.0
